services/api/requirements.txt
fastapi
uvicorn
nats-py
redis